rename `chat` folder to `copilot`
