+++ src/copilot/chatAgentRequestHandler.ts
